4 tab as common components and code cleanup
