+++ tvaultuiv2/src/views/private/safe/components/Permissions/index.js
@@ -1,16 +1,8 @@
 /* eslint-disable react/jsx-curly-newline */
-/* eslint-disable react/jsx-props-no-spreading */
 import React, { useState, useEffect, useCallback } from 'react';
 import styled, { css } from 'styled-components';
-import { makeStyles } from '@material-ui/core/styles';
-import Tab from '@material-ui/core/Tab';
 import PropTypes from 'prop-types';
-import AppBar from '@material-ui/core/AppBar';
-import Tabs from '@material-ui/core/Tabs';
 import ComponentError from '../../../../../errorBoundaries/ComponentError/component-error';
-import NamedButton from '../../../../../components/NamedButton';
-import permissionPlusIcon from '../../../../../assets/permission-plus.svg';
-import mediaBreakpoints from '../../../../../breakpoints';
 import User from './components/User';
 import Groups from './components/Groups';
 import apiService from '../../apiService';
@@ -18,104 +10,9 @@ import LoaderSpinner from '../../../../../components/Loaders/LoaderSpinner';
 import AppRoles from './components/AppRoles';
 import SnackbarComponent from '../../../../../components/Snackbar';
 import AwsApplications from './components/AwsApplications';
-
-const { belowLarge } = mediaBreakpoints;
-
-const TabPanelWrapper = styled.div`
-  height: 90%;
-`;
-
-function TabPanel(props) {
-  const { children, value, index } = props;
-  return (
-    <TabPanelWrapper
-      role="tabpanel"
-      hidden={value !== index}
-      id={`scrollable-prevent-tabpanel-${index}`}
-      aria-labelledby={`scrollable-prevent-tab-${index}`}
-    >
-      {children}
-    </TabPanelWrapper>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.number.isRequired,
-  value: PropTypes.number.isRequired,
-};
-
-TabPanel.defaultProps = {
-  children: <div />,
-};
-
-function a11yProps(index) {
-  return {
-    id: `scrollable-prevent-tab-${index}`,
-    'aria-controls': `scrollable-prevent-tabpanel-${index}`,
-  };
-}
-
-const TabWrapper = styled.div`
-  height: calc(100% - 3.8rem);
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  .MuiAppBar-colorPrimary {
-    background-color: inherit;
-  }
-  .MuiPaper-elevation4 {
-    box-shadow: none;
-  }
-  .MuiTabs-root {
-    min-height: unset;
-  }
-  .MuiTab-textColorInherit.Mui-selected {
-    background-color: ${(props) => props.theme.palette.secondary.main};
-    color: ${(props) => props.theme.palette.secondary.contrastText};
-  }
-  .MuiTab-textColorInherit {
-    color: #e8e8e8;
-    background-color: #20232e;
-    min-width: auto;
-    padding: 8.5px 20px;
-    margin-right: 0.5rem;
-    font-size: 1.4rem;
-    min-height: 3.65rem;
-    ${belowLarge} {
-      margin-right: 1.2rem;
-    }
-  }
-  .PrivateTabIndicator-colorSecondary-15 {
-    background-color: inherit;
-  }
-`;
-
-const CountPlusWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 2rem;
-  height: 3.8rem;
-`;
-const CountSpan = styled.div`
-  color: #5e627c;
-  font-size: 1.3rem;
-`;
-
-const customStyles = css`
-  display: flex;
-  ${belowLarge} {
-    display: none;
-  }
-`;
-
-const customMobileStyles = css`
-  display: none;
-  ${belowLarge} {
-    display: flex;
-  }
-`;
+import PermissionsTabs from '../../../../../components/PermissionTabs';
+import TabPanel from '../../../../../components/TabPanel';
+import { TabWrapper } from '../../../../../styles/GlobalStyles';
 
 const PermissionTabsWrapper = styled('div')`
   height: calc(100% - 3.7rem);
@@ -130,18 +27,8 @@ const customStyle = css`
   top: 50%;
 `;
 
-const useStyles = makeStyles(() => ({
-  appBar: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    height: '3.7rem',
-  },
-}));
-
 const Permissions = (props) => {
   const { safeDetail, refresh } = props;
-  const classes = useStyles();
   const [value, setValue] = useState(0);
   const [newPermission, setNewPermission] = useState(false);
   const [newGroup, setNewGroup] = useState(false);
@@ -150,26 +37,11 @@ const Permissions = (props) => {
   const [count, setCount] = useState(0);
   const [safeData, setSafeData] = useState({ response: {}, error: '' });
   const [responseType, setResponseType] = useState(0);
-  const [selectedTab, setSelectedTab] = useState('Permissions');
   const [toastResponse, setToastResponse] = useState(null);
   const [toastMessage, setToastMessage] = useState('');
 
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-    setCount(0);
-    if (newValue === 0) {
-      setSelectedTab('Permission');
-    } else if (newValue === 1) {
-      setSelectedTab('Group');
-    } else if (newValue === 2) {
-      setSelectedTab('AWS Application');
-    } else if (newValue === 3) {
-      setSelectedTab('App Role');
-    }
-  };
-
   useEffect(() => {
-    if (safeData?.response && Object.keys(safeData?.response).length !== 0) {
+    if (safeData?.response && Object.keys(safeData.response).length !== 0) {
       setCount(0);
       if (value === 0) {
         if (safeData.response.users) {
@@ -243,39 +115,20 @@ const Permissions = (props) => {
     setToastMessage(message);
   };
 
+  const onTabChange = (newValue) => {
+    setValue(newValue);
+  };
+
   return (
     <ComponentError>
       <>
-        <CountPlusWrapper>
-          <CountSpan color="#5e627c">{`${count} ${selectedTab}s`}</CountSpan>
-          <NamedButton
-            customStyle={customMobileStyles}
-            label={`Add ${selectedTab}`}
-            iconSrc={permissionPlusIcon}
-            onClick={() => onAddLabelBtnClicked()}
-          />
-        </CountPlusWrapper>
         <TabWrapper>
-          <AppBar position="static" className={classes.appBar}>
-            <Tabs
-              value={value}
-              onChange={handleChange}
-              variant="scrollable"
-              scrollButtons="off"
-              aria-label="scrollable prevent tabs example"
-            >
-              <Tab label="Users" {...a11yProps(0)} />
-              <Tab label="Groups" {...a11yProps(1)} />
-              <Tab label="AWS Applications" {...a11yProps(2)} />
-              <Tab label="App Roles" {...a11yProps(3)} />
-            </Tabs>
-            <NamedButton
-              customStyle={customStyles}
-              label={`Add ${selectedTab}`}
-              iconSrc={permissionPlusIcon}
-              onClick={() => onAddLabelBtnClicked()}
-            />
-          </AppBar>
+          <PermissionsTabs
+            count={count}
+            value={value}
+            onTabChange={onTabChange}
+            onAddLabelBtnClicked={onAddLabelBtnClicked}
+          />
           {responseType === 0 && <LoaderSpinner customStyle={customStyle} />}
           <PermissionTabsWrapper>
             {' '}
